requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
elasticsearch==8.13.0
sentence-transformers==2.7.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.1
websockets==12.0
aiofiles==23.2.1
